refactor: cleanup sqrt tests
