+++ .github/workflows/main_workflow.yml
@@ -33,5 +33,11 @@ jobs:
       - name: Checkout Code
         uses: actions/checkout@v3
 
+      - name: Setup Java
+        uses: actions/setup-java@v3
+        with:
+          distribution: 'corretto'
+          java-version: '17'
+
       - name: Unit Tests
-        run: ./gradlew test
+        run: ./gradlew test --info --stacktrace
