+++ README.md
@@ -9,13 +9,15 @@
 ExactFraction is an implementation of the [Number](https://docs.oracle.com/javase/8/docs/api/java/lang/Number.html) class.
 It was inspired by the [BigDecimal](https://docs.oracle.com/javase/8/docs/api/java/math/BigDecimal.html) class, but fills a slightly different purpose.
 An ExactFraction is an exact representation of a rational number.
-It stores the number as a pair of BigInteger's, representing the numerator and denominator, which means that infinite rational numbers (i.e. 5/9) can be stored without rounding.
-Therefore, operations with ExactFraction's can be performed without losing precision due to rounding.
+It stores the number as a pair of BigIntegers, representing the numerator and denominator, which means that infinitely repeating rational numbers (i.e. 5/9) can be stored without rounding.
+Therefore, operations with ExactFractions can be performed without losing precision due to rounding.
 
 ### Irrational
 
 Though the exact value of an irrational number can't be stored, the value is represented as a collection of rational numbers related to the irrational.
 This allows numbers to be multiplied, divided, and simplified before computing the final value, which allows for greater precision when getting the value.
+
+## Expression Types
 
 ### Term
 
@@ -71,11 +73,14 @@ TOKEN=GITHUB_PAT
 However, this configuration may not allow you to build through an IDE.
 If you have values set in both gradle.properties and in the environment, the values in gradle.properties will be used.
 
+**Do not commit your access token.**
+
 See [here](https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-gradle-registry#using-a-published-package) for more information on importing GitHub packages.
 
 ## Testing
 
 Unit tests are written using the [Kotlin test](https://kotlinlang.org/api/latest/kotlin.test/) framework.
+Mocking is done using the [MockK library](https://mockk.io/).
 Tests must be written for all logic in the package.
 
 Tests can be run using an IDE, or with the following command:
@@ -83,7 +88,7 @@ Tests can be run using an IDE, or with the following command:
 
 ### Resources
 
-When writing tests for irrationals, it can be useful to use a high-precision calculator, such as [this one](https://www.mathsisfun.com/scientific-calculator.html), to calculate expected values.
+When writing tests for irrationals, it can be useful to use a high-precision calculator, such as [this](https://www.mathsisfun.com/scientific-calculator.html) or [this](https://www.mathsisfun.com/calculator-precision.html), to calculate expected values.
 
 ## Linting
