+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/exactfraction/constructorTests.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.exactnumbers.exactfraction
 
-import assertDivByZero
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import java.math.BigInteger
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
