+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/irrationals/common/IrrationalOperatorsTest.kt
@@ -1,10 +1,10 @@
 package xyz.lbres.exactnumbers.irrationals.common
 
-import assertDivByZero
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import xyz.lbres.expressions.term.Term
 import kotlin.test.Test
 import kotlin.test.assertEquals
