+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/irrationals/log/constructorTests.kt
@@ -1,7 +1,7 @@
 package xyz.lbres.exactnumbers.irrationals.log
 
-import assertDivByZero
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import java.math.BigInteger
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
