+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/irrationals/sqrt/binaryOpTests.kt
@@ -1,11 +1,13 @@
 package xyz.lbres.exactnumbers.irrationals.sqrt
 
-import assertDivByZero
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import xyz.lbres.expressions.term.Term
 import kotlin.test.assertEquals
+import kotlin.test.assertNotEquals
+import kotlin.test.assertTrue
 
 private val one = ExactFraction.ONE
 
@@ -101,4 +103,65 @@ fun runDivTests() {
     sqrt2 = Sqrt(ExactFraction.TWO)
     expected = Term(one, listOf(Sqrt(ExactFraction.HALF), Sqrt(ExactFraction.HALF)))
     assertEquals(expected, sqrt1 / sqrt2)
+}
+
+fun runCompareToTests() {
+    // equal
+    var sqrt1 = Sqrt.ZERO
+    assertEquals(0, sqrt1.compareTo(sqrt1))
+
+    sqrt1 = Sqrt(ExactFraction(9, 11))
+    assertEquals(0, sqrt1.compareTo(sqrt1))
+
+    // not equal
+    sqrt1 = Sqrt.ZERO
+    var sqrt2 = Sqrt.ONE
+    assertTrue(sqrt1 < sqrt2)
+    assertTrue(sqrt2 > sqrt1)
+
+    sqrt1 = Sqrt(3)
+    sqrt2 = Sqrt(9)
+    assertTrue(sqrt1 < sqrt2)
+    assertTrue(sqrt2 > sqrt1)
+
+    sqrt1 = Sqrt(ExactFraction(15, 26))
+    sqrt2 = Sqrt(ExactFraction(26, 15))
+    assertTrue(sqrt1 < sqrt2)
+    assertTrue(sqrt2 > sqrt1)
+}
+
+fun runEqualsTests() {
+    // equal
+    var sqrt1 = Sqrt.ZERO
+    assertEquals(sqrt1, sqrt1)
+
+    sqrt1 = Sqrt(6)
+    assertEquals(sqrt1, sqrt1)
+
+    sqrt1 = Sqrt(ExactFraction(9, 400))
+    assertEquals(sqrt1, sqrt1)
+
+    sqrt1 = Sqrt(ExactFraction(400, 9))
+    assertEquals(sqrt1, sqrt1)
+
+    // not equal
+    sqrt1 = Sqrt.ZERO
+    var sqrt2 = Sqrt.ONE
+    assertNotEquals(sqrt1, sqrt2)
+    assertNotEquals(sqrt2, sqrt1)
+
+    sqrt1 = Sqrt(2)
+    sqrt2 = Sqrt(ExactFraction.HALF)
+    assertNotEquals(sqrt1, sqrt2)
+    assertNotEquals(sqrt2, sqrt1)
+
+    sqrt1 = Sqrt(ExactFraction(9, 25))
+    sqrt2 = Sqrt(ExactFraction.NINE)
+    assertNotEquals(sqrt1, sqrt2)
+    assertNotEquals(sqrt2, sqrt1)
+
+    sqrt1 = Sqrt(ExactFraction(103, 422))
+    sqrt2 = Sqrt(ExactFraction(90, 37))
+    assertNotEquals(sqrt1, sqrt2)
+    assertNotEquals(sqrt2, sqrt1)
 }
